+++ back/src/__tests__/seen/create.test.ts
@@ -1,22 +1,26 @@
 import supertest from 'supertest'
 import { expect, describe, beforeAll, afterAll, it, afterEach } from 'vitest'
-import { Category, Resource } from '@prisma/client'
+import { Category, Resource, User } from '@prisma/client'
 import { server, testUserData } from '../globalSetup'
-import { authToken } from '../setup'
 import { prisma } from '../../prisma/client'
 import { pathRoot } from '../../routes/routes'
 import { resourceTestData } from '../mocks/resources'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
+import { authToken } from '../mocks/ssoServer'
 
 let testResource: Resource
 const uri = `${pathRoot.v1.seen}/`
+let user: User | null
 beforeAll(async () => {
   const testCategory = (await prisma.category.findUnique({
     where: { slug: 'testing' },
   })) as Category
+  user = await prisma.user.findFirst({
+    where: { name: testUserData.user.name },
+  })
   const testResourceData = {
     ...resourceTestData[0],
-    user: { connect: { dni: testUserData.user.dni } },
+    user: { connect: { id: user?.id } },
     category: { connect: { id: testCategory.id } },
   }
   testResource = await prisma.resource.create({
@@ -27,7 +31,7 @@ beforeAll(async () => {
 afterAll(async () => {
   await prisma.viewedResource.deleteMany({ where: {} })
   await prisma.resource.deleteMany({
-    where: { user: { dni: testUserData.user.dni } },
+    where: { user: { id: user?.id } },
   })
 })
 
@@ -38,10 +42,10 @@ describe('Testing viewed resource creation endpoint', () => {
   it('should mark a resource as viewed', async () => {
     const response = await supertest(server)
       .post(`${pathRoot.v1.seen}/${testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     const viewedResources = await prisma.viewedResource.findMany({
       where: {
-        user: { dni: testUserData.admin.dni },
+        user: { name: testUserData.admin.name },
         resourceId: testResource.id,
       },
     })
@@ -52,15 +56,15 @@ describe('Testing viewed resource creation endpoint', () => {
   it('should not create duplicate entries for viewed resources', async () => {
     const response = await supertest(server)
       .post(`${uri + testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(response.statusCode).toBe(204)
     const secondResponse = await supertest(server)
       .post(`${uri + testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(secondResponse.statusCode).toBe(204)
     const viewedResources = await prisma.viewedResource.findMany({
       where: {
-        user: { dni: testUserData.admin.dni },
+        user: { name: testUserData.admin.name },
         resourceId: testResource.id,
       },
     })
@@ -70,7 +74,7 @@ describe('Testing viewed resource creation endpoint', () => {
     it('Should response status code 400 if no valid cuid is provided', async () => {
       const response = await supertest(server)
         .post(`${uri}abc`)
-        .set('Cookie', authToken.admin)
+        .set('Cookie', [`authToken=${authToken.admin}`])
       expect(response.status).toBe(400)
       expect(response.body.message[0]).toStrictEqual({
         code: 'invalid_string',
@@ -92,7 +96,7 @@ describe('Testing viewed resource creation endpoint', () => {
     it('should response status code 404 if resource does not exist', async () => {
       const response = await supertest(server)
         .post(`${uri}clnjyhw7t000008ju4ozndeqi`)
-        .set('Cookie', authToken.admin)
+        .set('Cookie', [`authToken=${authToken.admin}`])
 
       expect(response.statusCode).toBe(404)
       expect(response.body.message).toBe('Resource not found')